// server/src/types.ts
type typeLawReview = {
  region: string;
  client: string;
  usage: string;
  approval: string;
  number_of_households: string;
  land_area: string;
  floor_area: string;
  parking_area: string;
  number_of_floors: string;
  construct: string;
  heating: string;
  education: string;
  maintenance: string;
  rivers: string;
};

export type { typeLawReview };